requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2023.1
crispy-bootstrap5==2023.10
cryptography==41.0.7
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-filter==23.5
django-froala-editor==4.1.4
django-import-export==3.3.5
django-js-asset==2.2.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
freetype-py==2.3.0
html5lib==1.1
idna==3.6
lxml==4.9.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
Pillow==10.1.0
psycopg2==2.9.9
pycairo==1.25.1
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pypdf==3.17.4
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.8
requests==2.31.0
rlPyCairo==0.3.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tablib==3.5.0
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.13
xlrd==2.0.1
xlwt==1.3.0
